requirements.txt
absl-py==1.3.0
albumentations==1.3.0
asttokens==2.0.8
av==9.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blis==0.7.8
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cryptography==37.0.4
cycler==0.11.0
cymem==2.0.6
dataclasses==0.8
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.5.0
executing==1.0.0
filelock==3.8.0
fonttools==4.37.2
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20221221
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.9.1
idna==3.4
imageio==2.26.0
imbalanced-learn==0.9.0
importlib-metadata==4.11.4
iopath==0.1.10
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
munkres==1.1.4
murmurhash==1.0.8
mypy==0.971
mypy-extensions==0.4.3
networkx==3.0
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
orjson==3.8.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
ply==3.11
portalocker==2.7.0
preshed==3.0.7
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.20.2
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
sacremoses==0.0.43
SciencePlots==2.1.1
scikit-image==0.20.0
scikit-learn==1.0.2
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.9.8
setproctitle==1.3.2
setuptools==65.3.0
shellingham==1.5.0
shortuuid==1.0.9
sip==6.6.2
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stack-data==0.5.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
thinc==8.1.0
threadpoolctl==3.1.0
tifffile==2023.3.15
tokenizers==0.11.4
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchinfo==1.7.0
torchvision==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.22.1
typer==0.4.2
typing_extensions==4.3.0
ujson==5.6.0
unicodedata2==14.0.0
urllib3==1.26.11
wandb==0.13.3
wasabi==0.10.0
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
xgboost==1.6.2
yacs==0.1.8
zipp==3.8.1
